Stergerea de fisiere repetitive si adaugarea mai multor endpoint-uri pe partea de profile

+++ AgoraCampus/src/main/java/com/agora/agoracampus/profile/core/mapper/ProfileMapper.java
@@ -14,7 +14,7 @@ public class ProfileMapper {
         public ProfileResponse toResponse(Profile profile) {
             return new ProfileResponse(
                     profile.getId(),
-                    profile.getId().longValue(),
+                    profile.getId(),
                     profile.getHeadline(),
                     profile.getDescription(),
                     profile.getLocation(),
